backend/requirements.txt
fastapi
uvicorn
sympy
numpy
requests
python-dotenv